scripts: trim trailing blank lines from npmjs.org
The readme strings stored in npmjs.org correctly end in a terminating
newline. Unfortunately, the `jq` utility _also_ terminates its output
with a newline, so when we pipe out the data it ends up added an _extra_
blank line at the end of the file (ends in `\n\n` instead of `\n`).